doing islandr in free time cuz why not
fixed obstacles "not rlly there but still being rendered" due to the obstacle previously recognized in map packet but since in server, when broken, removes the obstacle from the list, disregards updating it in client.
made a duplicate of the obstacle list which doesnt have remove operations and that's what is being sent to the client

for ghost players, i used the killfeed ("announcement") packet for this, where i now send the killed person's position info, and manually update it in the client, so the client updates info on that. (rest of the entities dont die lol so doesnt matter)